+++ src/testHarness/OrderBook.java
@@ -1,5 +1,6 @@
 package testHarness;
 
+import java.math.BigDecimal;
 import java.sql.Timestamp;
 import java.util.Iterator;
 
@@ -28,7 +29,7 @@ public abstract class OrderBook {
 	 * @param time Time at which the order was placed.
 	 * @return A reference to the buy order.
 	 */
-	public abstract BuyOrder buy(int volume, int price, Timestamp time);
+	public abstract BuyOrder buy(int volume, BigDecimal price, Timestamp time);
 	
 	/**
 	 * Places an ask (sell) order into the order book.
@@ -37,7 +38,7 @@ public abstract class OrderBook {
 	 * @param time Time at which the order was placed.
 	 * @return A reference to the sell order.
 	 */
-	public abstract SellOrder sell(int volume, int price, Timestamp time);
+	public abstract SellOrder sell(int volume, BigDecimal price, Timestamp time);
 	
 	/**
 	 * @return An iterator of all the bids in the order book.
